+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.9.14] - unreleased
 ### Added
+- GHORG_NO_DIR_SIZE flag to turn off directory size output which is now enabled by default
 ### Changed
 ### Deprecated
 ### Removed